% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyt.anova.R
\name{cyt.anova}
\alias{cyt.anova}
\title{ANOVA analysis on all continuous variables within the data.}
\usage{
cyt.anova(x.df)
}
\arguments{
\item{x.df}{A matrix or data frame consisting of continuous and categorical variables.}
}
\value{
Prints the p-values of comparisons conducted using the Tukey test of the ANOVA model.
}
\description{
This function produces and prints list of p-values obtained from Tukey comparisons. It assumes that
the first two columns are categorical variables whether it is group name or treatment name and uses
those as predictors. The rest of the columns in the data set are assumed to be continuous variables to
be used as the outcomes.
}
\examples{
data("cytdata.df")
cyt.anova(cytdata.df[,c(2:4,5:6)])
}
