% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyt.bp.R
\name{cyt.bp}
\alias{cyt.bp}
\title{Generating a PDF file to show the measured value by column of the data frame.}
\usage{
cyt.bp(data, Title, bin.size = 25, mfRow = c(1, 1), yLim = NULL)
}
\arguments{
\item{data}{A matrix or data frame containing the raw data to be plotted.}

\item{Title}{A string representing the name of the PDF file.}

\item{bin.size}{The maximum number of box plots to display on a single page.}

\item{mfRow}{A numeric vector of length two specifying the layout (rows and columns) for the plots on each page.}

\item{yLim}{An optional numeric vector defining the y-axis limits for the plots.}
}
\value{
A PDF file containing the box plots of the continuous variables.
}
\description{
This function creates a PDF file containing box plots for the continuous variables in the provided data.
If the number of columns in \code{data} exceeds \code{bin.size}, the function splits the plots across multiple pages.
}
\examples{
\dontrun{
  # Generating boxplots to check for outliers for raw values:
  cyt.bp(data.df[,-c(1:4)], Title = "Boxplot.byCytokine.Raw.pdf")

  # Generating boxplots to check for outliers for log2 values:
  cyt.bp(log2(data.df[,-c(1:4)]), Title = "Boxplot.byCytokine.log2.pdf")
}
}
