% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyt.bp2.R
\name{cyt.bp2}
\alias{cyt.bp2}
\title{Boxplot Function Enhanced for Specific Group Comparisons}
\usage{
cyt.bp2(x.df, Title, mfRow = c(1, 1), scale = NULL, yLim = NULL)
}
\arguments{
\item{x.df}{A matrix or data frame of raw data.}

\item{Title}{The title of the PDF file.}

\item{scale}{Transformation option for continuous variables. Options are NULL (default) and "log2".}

\item{yLim}{A range for the y-axis to be plotted.}
}
\value{
A PDF file containing boxplots of the data.
}
\description{
Boxplot Function Enhanced for Specific Group Comparisons
}
\examples{
data.df = cytodata[,-c(1,4)]
cyt.bp2(data.df, Title = "boxplot2.test2.pdf", scale = "log2")
}
