% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyt.dualflashplot.R
\name{cyt.dualflashplot}
\alias{cyt.dualflashplot}
\title{Dual flash plot.}
\usage{
cyt.dualflashplot(
  data,
  group_var,
  group1,
  group2,
  ssmd_thresh = 1,
  log2fc_thresh = 1,
  top_labels = 15
)
}
\arguments{
\item{data}{A data frame to be entered.}

\item{group_var}{Grouping column to be used from the data frame.}

\item{group1}{Name of the first group to be compared.}

\item{group2}{Name of the second group to be compared.}

\item{ssmd_thresh}{Threshold value for SSMD. Default set to 1.}

\item{log2fc_thresh}{Threshold value for log2 fold change. Default set to 1.}

\item{top_labels}{Number of labels to be print for top variables. Default set to 15.}
}
\value{
Prints dual flash plot.
}
\description{
Dual flash plot.
}
\examples{
cyt.dualflashplot(cytdata.df[,-c(1,3)], group_var = "Group", group1 = "ND", group2 = "T2D", ssmd_thresh = -0.2, log2fc_thresh = 1, top_labels = 10)
}
