% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyt.heatmap.R
\name{cyt.heatmap}
\alias{cyt.heatmap}
\title{Heat Map}
\usage{
cyt.heatmap(data, scale = NULL, annotation_col_name = NULL, title)
}
\arguments{
\item{data}{A data frame to be entered.}

\item{scale}{Log2 transformation for scaling. Default set to NULL.}

\item{annotation_col_name}{A string of a column name to be chosen for annotation. Default set to NULL.}

\item{title}{Title of the heatmap to be saved. Can save it as png or pdf.}
}
\value{
Prints a heatmap.
}
\description{
This function takes a data frame and subsets the numeric columns or variables to generate a heatmap.
}
\examples{
cyt.heatmap(cytodata[,-4], scale = "log2", annotation_col_name = "Group")
}
