% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyt.pca.R
\name{cyt.pca}
\alias{cyt.pca}
\title{Analyze data with Principal Component Analysis (PCA) for cytokines.}
\usage{
cyt.pca(
  data.df,
  colors = NULL,
  title,
  ellipse = FALSE,
  comp.num = 2,
  scale = NULL,
  pch.values = NULL,
  style = NULL
)
}
\arguments{
\item{data.df}{A data frame containing cytokine data. The first two columns are assumed to hold grouping information (e.g., group and treatment/stimulation).}

\item{colors}{A vector of colors corresponding to the groups. If set to NULL, a random palette (using rainbow) is generated based on the number of groups.}

\item{title}{A string specifying the  file name of the PDF file where the PCA plots will be saved.}

\item{ellipse}{Logical. If TRUE, a 95\% confidence ellipse is drawn on the PCA plot. Default is FALSE.}

\item{comp.num}{Numeric. The number of principal components to compute and display. Default is 2.}

\item{scale}{Character. Scaling option; if set to "log2", a log2 transformation is applied to the cytokine measurements (excluding the grouping columns). Default is NULL.}

\item{pch.values}{A vector of plotting characters to be used in the PCA plots. Default is NULL.}

\item{style}{Character. If set to "3d" or "3D" and \code{comp.num} equals 3, a 3D scatter plot is generated using the plot3D package. Default is NULL.}
}
\value{
A PDF file containing the PCA plots is generated and saved.
}
\description{
This function performs Principal Component Analysis (PCA) on cytokine data and generates several types of plots,
including:
\itemize{
\item 2D PCA plots using mixOmics's \code{plotIndiv} function,
\item 3D scatter plots (if \code{style} is "3d" or "3D" and \code{comp.num} is 3) via the plot3D package,
\item Scree plots showing both individual and cumulative explained variance,
\item Loadings plots, and
\item Biplots and correlation circle plots.
}
The function optionally applies a log2 transformation to the numeric data and handles analyses based on either treatment
or stimulation groups.
}
\examples{
data <- cytodata[,-c(1,4)]
data.df <- filter(data, Group != "ND" & Treatment != "Unstimulated")
data.df <- data.df[,-22]
cyt.pca(data.df, title = "PCA_Example_Analysis.pdf" ,colors = c("black", "red2"),
scale = "log2", comp.num = 3, pch.values = c(16,4), style = "3D")
cyt.pca(data.df, title = "PCA_Example_Analysis2.pdf" ,colors = c("black", "red2"),
scale = "log2", comp.num = 2, ellipse = TRUE, pch.values = c(16,4))
}
