% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyt.plsda.R
\name{cyt.plsda}
\alias{cyt.plsda}
\title{Analyze data with PLS-DA method.}
\usage{
cyt.plsda(
  data.df,
  colors = NULL,
  title,
  ellipse = FALSE,
  bg = FALSE,
  conf.mat = FALSE,
  var.num,
  cv.opt = NULL,
  fold.num = 5,
  scale = NULL,
  comp.num = 2,
  pch.values
)
}
\arguments{
\item{colors}{Vector of colors to be set, list of colors to be set equal to the number of groups or treatments. Default set to NULL and will result in random colors.}

\item{title}{Title of the PDF file to be saved.}

\item{ellipse}{If ellipse should be drawn in the figures or not. Default set to FALSE.}

\item{bg}{If prediction background should be drawn in the figures or not. Default set to FALSE.}

\item{conf.mat}{If confusion matrix of the classifications be printed at the end or not. Default set to FALSE.}

\item{var.num}{Number of variables to be used in PLS-DA.}

\item{cv.opt}{Option for "loocv" or "Mfold" cross-validation methods.}

\item{fold.num}{Number of folds if cv.opt equals "Mfold".}

\item{scale}{Option to transform data using log2. Default set to NULL.}

\item{comp.num}{Number of components to be calculated. Default set to 2.}

\item{pch.values}{A vector of integers specifying the plotting characters to be used in the plot.}

\item{x.df}{A matrix or data frame of variables.}
}
\value{
A PDF file consisting of classification figures, component figures including Variable of importance in projection (VIP) scores, and
classification of groups with VIP scores greater than 1.
}
\description{
This function takes a matrix or data frame of variables where the first two columns must consist of "Group"
and "Stimulation" or "Group" and "Treatment". The rest of columns will be cytokines that will be analyzed.
The function also uses log2 transformation of the data, so the data that is inputed, must be raw data without
any transformation.
}
\details{
This function takes a data frame assuming that the first two columns consists
of groups and stimulation of patients and using the rest of the columns as
components for the PLS-DA analysis.
}
\examples{
\dontrun{
cyt.plsda(x.df, title = "Example PLS-DA Analysis.pdf", bg = TRUE, conf.mat = TRUE, scale = "log2",
var.num = 25, cv.opt = "loocv", comp.num = 2, colors = c("black", "purple", "red2"),
pch.values = c(16,4,3))
}
}
