% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyt.ttest.R
\name{cyt.ttest}
\alias{cyt.ttest}
\title{Two Sample T-test Comparisons}
\usage{
cyt.ttest(x.df, scale = NULL)
}
\arguments{
\item{x.df}{A matrix or data frame containing continuous variables and categorical variables.}

\item{scale}{A character value specifying a transformation for continuous variables. Options are \code{NULL} (default) and \code{"log2"}.
When \code{scale = "log2"}, a log2 transformation is applied and a two-sample t-test is used; when \code{scale} is \code{NULL}, a Mann-Whitney U test is performed.}
}
\value{
A list of p-values from the statistical tests for each combination of continuous outcome and categorical predictor is returned.
}
\description{
This function performs pairwise comparisons between two groups for each combination of a categorical predictor (with exactly two levels)
and a continuous outcome variable. It first converts any character variables in \code{x.df} to factors and applies a log2 transformation to the continuous
variables if specified. Depending on the value of \code{scale}, the function conducts either a two-sample t-test or a Mann-Whitney U test and prints the resulting
p-values. An error is thrown if a categorical variable does not have exactly two levels.
}
\examples{
\dontrun{
  data.df <- cytodata[,-c(1,4)]
  data.df <- filter(data.df, Group != "ND", Treatment != "Unstimulated")
  # Two sample T-test
  cyt.ttests(data.df[, c(1,2, 5:6)], scale = "log2")
  # Mann Whitney U Test
  cyt.ttests(data.df[, c(1,2, 5:6)])
}
}
