% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyt.ttest.R
\name{cyt.ttests}
\alias{cyt.ttests}
\title{Two Sample T-test Comparisons}
\usage{
cyt.ttests(x.df, scale = NULL)
}
\arguments{
\item{x.df}{A matrix or data frame consisting of continuous and categorical variables.}

\item{scale}{Transformation option for continuous variables. Options are NULL (default) and "log2".
When scale = "log2". the function will use two sample t-test, when scale = NULL, the function will use Mann-Whitney U test.}
}
\value{
Prints the p-values of comparisons conducted.
}
\description{
Two Sample T-test Comparisons
}
\examples{
data.df = cytdata.df[,-c(1,4)]
data.df = filter(data.df, Group != "ND", Treatment != "Unstimulated")
cyt.ttests(data.df, scale = "log2")
cyt.ttests(data.df)

}
