% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyt.volc.R
\name{cyt.volc}
\alias{cyt.volc}
\title{Volcano Plot}
\usage{
cyt.volc(
  data,
  group_col,
  cond1 = NULL,
  cond2 = NULL,
  fold_change_thresh = 2,
  p_value_thresh = 0.05,
  top_labels = 10
)
}
\arguments{
\item{group_col}{Column used for comparison (i.e. group, treatment, or stimulation).}

\item{cond1}{A string of a group name to be chosen. Default set to NULL.}

\item{cond2}{A string of a group name to be chosen. Default set to NULL.}

\item{fold_change_thresh}{Threshold value for fold change. Default set to 2.}

\item{p_value_thresh}{Threshold value for p-value. Default set to 0.05.}

\item{top_labels}{Number of labels to be print for top variables. Default set to 10}

\item{x.df}{A matrix or data frame}
}
\value{
Prints volcano plot.
}
\description{
This function takes a matrix or data frame and subsets the numeric columns or variables
to compare against each other based on a selected column of choice that defines a category
to draw a volcano plot.
}
\note{
If cond1 and cond2 are set to null, by default the function creates combinations of pairs
to be used for comparisons.
}
\examples{
cyt.volc(cytdata.df,group_col = "Group")
cyt.volc(cytdata.df, group_col = "Group", fold_change_thresh = 2, top_labels = 15)
}
