% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExampleData1}
\alias{ExampleData1}
\title{Example Cytokine Profiling Data 1.}
\format{
A data frame with 297 rows and 29 columns:
\describe{
\item{Group}{Group assigned to the subjects.}
\item{Treatment}{Treatment received by subjects.}
\item{Time}{Time point of the measurement.}
\item{IL-17F}{Observed concentration of IL-17F cytokine.}
\item{GM-CSF}{Observed concentration of GM-CSF cytokine.}
\item{IFN-G}{Observed concentration of IFN-G cytokine.}
\item{IL-10}{Observed concentration of IL-10 cytokine.}
\item{CCL-20/MIP-3A}{Observed concentration of CCL-20/MIP-3A cytokine.}
\item{IL-12/P70}{Observed concentration of IL-12/P70 cytokine.}
\item{IL-13}{Observed concentration of IL-13 cytokine.}
\item{IL-15}{Observed concentration of IL-15 cytokine.}
\item{IL-17A}{Observed concentration of IL-17A cytokine.}
\item{IL-22}{Observed concentration of IL-22 cytokine.}
\item{IL-9}{Observed concentration of IL-9 cytokine.}
\item{IL-1B}{Observed concentration of IL-1B cytokine.}
\item{IL-33}{Observed concentration of IL-33 cytokine.}
\item{IL-2}{Observed concentration of IL-2 cytokine.}
\item{IL-21}{Observed concentration of IL-21 cytokine.}
\item{IL-4}{Observed concentration of IL-4 cytokine.}
\item{IL-23}{Observed concentration of IL-23 cytokine.}
\item{IL-5}{Observed concentration of IL-5 cytokine.}
\item{IL-6}{Observed concentration of IL-6 cytokine.}
\item{IL-17E/IL-25}{Observed concentration of IL-17E/IL-25 cytokine.}
\item{IL-27}{Observed concentration of IL-27 cytokine.}
\item{IL-31}{Observed concentration of IL-31 cytokine.}
\item{TNF-A}{Observed concentration of TNF-A cytokine.}
\item{TNF-B}{Observed concentration of TNF-B cytokine.}
\item{IL-28A}{Observed concentration of IL-28A cytokine.}
}
}
\source{
Example data compiled for cytokine profiling.
}
\usage{
ExampleData1
}
\description{
Contains observed concentrations of cytokines and their respective treatment
and groups, derived from:
}
\examples{
data(ExampleData1)
}
\references{
Pugh GH, Fouladvand S, SantaCruz-Calvo S, Agrawal M, Zhang XD, Chen J,
Kern PA, Nikolajczyk BS.
T cells dominate peripheral inflammation in a cross-sectional analysis of
obesity-associated diabetes. \emph{Obesity (Silver Spring)}. 2022;30(10):
1983–1994. doi:10.1002/oby.23528.
}
\keyword{datasets}
