% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExampleData4}
\alias{ExampleData4}
\title{Example Cytokine Profiling Data 4.}
\format{
A data frame with 64 rows and 14 columns:
\describe{
\item{Group}{Group assigned to the subjects.}
\item{Treatment}{Treatment received by subjects.}
\item{IL-17F}{Observed concentration of IL-17F cytokine.}
\item{GM-CSF}{Observed concentration of GM-CSF cytokine.}
\item{IFN-G}{Observed concentration of IFN-G cytokine.}
\item{IL-10}{Observed concentration of IL-10 cytokine.}
\item{CCL-20}{Observed concentration of CCL-20 cytokine.}
\item{IL-12}{Observed concentration of IL-12 cytokine.}
\item{IL-13}{Observed concentration of IL-13 cytokine.}
\item{IL-17A}{Observed concentration of IL-17A cytokine.}
\item{IL-22}{Observed concentration of IL-22 cytokine.}
\item{IL-9}{Observed concentration of IL-9 cytokine.}
\item{IL-2}{Observed concentration of IL-2 cytokine.}
\item{IL-21}{Observed concentration of IL-21 cytokine.}
\item{IL-4}{Observed concentration of IL-4 cytokine.}
\item{IL-23}{Observed concentration of IL-23 cytokine.}
\item{IL-5}{Observed concentration of IL-5 cytokine.}
\item{IL-6}{Observed concentration of IL-6 cytokine.}
\item{TNF-A}{Observed concentration of TNF-A cytokine.}
\item{TNF-B}{Observed concentration of TNF-B cytokine.}
}
}
\source{
Example data compiled for cytokine profiling.
}
\usage{
ExampleData4
}
\description{
Contains observed concentrations of cytokines and their respective treatment
and groups, derived from:
}
\examples{
data(ExampleData4)
}
\references{
SantaCruz-Calvo, S., Saraswat, S., Kalantar, G. H., Zukowski, E.,
Marszalkowski, H., Javidan, A., Gholamrezaeinejad, F., Bharath, L. P.,
Kern, P. A., Zhang, X. D., & Nikolajczyk, B. S. (2024).
A unique inflammaging profile generated by T cells from people with obesity
is metformin resistant. \emph{GeroScience}, 10.1007/s11357-024-01441-4.
Advance online publication. https://doi.org/10.1007/s11357-024-01441-4
}
\keyword{datasets}
