% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyt_graphs.R
\name{plot_group_distribution}
\alias{plot_group_distribution}
\title{Distribution of the groups within the continuous variable.}
\usage{
plot_group_distribution(x.df, continuous_var, group_var)
}
\arguments{
\item{x.df}{A matrix or data frame consisting of continuous and categorical variables.}

\item{cont_var}{Continuous variable.}

\item{group}{Column name to be color coded by.}
}
\value{
Prints the histograms.
}
\description{
Distribution of the groups within the continuous variable.
}
