% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyt.bp2.R
\name{cyt.bp2}
\alias{cyt.bp2}
\title{Boxplot Function Enhanced for Specific Group Comparisons}
\usage{
cyt.bp2(data, Title, mfRow = c(1, 1), scale = NULL, yLim = NULL)
}
\arguments{
\item{data}{A matrix or data frame of raw data.}

\item{Title}{A string representing the title of the PDF file.}

\item{mfRow}{A numeric vector of length two specifying the layout (rows and columns) for the plots on each page. Defaults to c(1,1).}

\item{scale}{Transformation option for continuous variables. Options are NULL (default) and "log2". When set to "log2", numeric columns are transformed using the log2 function.}

\item{yLim}{An optional numeric vector defining the y-axis limits for the plots.}
}
\value{
A PDF file containing the boxplots.
}
\description{
This function generates a PDF file containing boxplots for each combination of numeric and factor variables in the provided data.
It first converts any character columns to factors and checks that the data contains at least one numeric and one factor column.
If the scale argument is set to "log2", all numeric columns are log2-transformed.
The function then creates boxplots using ggplot2 for each numeric variable grouped by each factor variable.
}
\examples{
\dontrun{
data.df <- cytodata[, -c(1, 4)]
cyt.bp2(data.df, Title = "boxplot2.test2.pdf", scale = "log2")
}
}
