% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyt.errbp.R
\name{cyt.errbp}
\alias{cyt.errbp}
\title{Error-bar plot for comparison.}
\usage{
cyt.errbp(
  data,
  pLab = TRUE,
  esLab = TRUE,
  classSymbol = TRUE,
  xlab = "",
  ylab = "",
  main = ""
)
}
\arguments{
\item{data}{A data frame containing the following columns for each group:
\itemize{
\item \code{name}: Group names.
\item \code{center}: Mean or median values.
\item \code{spread}: Standard deviation, MAD, or standard error.
\item \code{p.value}: P-value for the comparison.
\item \code{effect.size}: Effect size based on SSMD.
}
Note: The first row of \code{center.df} must correspond to the baseline group.}

\item{pLab}{Logical. Whether to label the p-values on the plot. Default is \code{TRUE}.}

\item{esLab}{Logical. Whether to label the effect sizes on the plot. Default is \code{TRUE}.}

\item{classSymbol}{Logical. Whether to use symbolic notation for significance and effect size. Default is \code{TRUE}.}

\item{xlab}{Character. Label for the x-axis.}

\item{ylab}{Character. Label for the y-axis.}

\item{main}{Character. Title of the graph.}
}
\value{
An error-bar plot is produced.
}
\description{
This function draws an error-bar plot for comparing groups to a baseline group. It creates a barplot of
the central tendency (mean or median) and overlays error bars representing the spread (e.g., standard deviation,
MAD, or standard error). Optionally, p-value and effect size labels (based on SSMD) are added, either as symbols
or numeric values.
}
\examples{
\dontrun{
cytokine.mat <- cytodata[, -c(1:4)] # Extracting all cytokines to be stored in one object
cytokineNames <- colnames(cytokine.mat) # Extracting the cytokine names
nCytokine <- length(cytokineNames) # Obtaining the total number of cytokines
results <- cyt.skku(cytodata[, -c(1, 4)],
  printResLog = TRUE,
  group.cols = c("Group", "Treatment")
) # Extracting values
pdf("barErrorPlot.pdf")
par(mfrow = c(2, 2), mar = c(8.1, 4.1, 4.1, 2.1))
for (k in 1:nCytokine) {
  center.df <- data.frame("name" = rownames(results[, , k]), results[, , k])
  cyt.errbp(center.df,
    pLab = FALSE, esLab = FALSE, classSymbol = TRUE,
    ylab = "Concentration in log2 scale", main = cytokineNames[k]
  )
}
dev.off()
}

}
