% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyt.pca.R
\name{cyt.pca}
\alias{cyt.pca}
\title{Analyze data with Principal Component Analysis (PCA) for cytokines.}
\usage{
cyt.pca(
  data,
  group.col = NULL,
  trt.col = NULL,
  colors = NULL,
  title,
  ellipse = FALSE,
  comp.num = 2,
  scale = NULL,
  pch.values = NULL,
  style = NULL
)
}
\arguments{
\item{data}{A data frame containing cytokine data. It should include at least one column representing grouping information and optionally a second column representing treatment or stimulation.}

\item{group.col}{Character. The name of the column containing the grouping information. If not specified and \code{trt.col} is provided, the treatment column will be used as the grouping variable.}

\item{trt.col}{Character. The name of the column containing the treatment (or stimulation) information. If not specified and \code{group.col} is provided, the grouping column will be used as the treatment variable.}

\item{colors}{A vector of colors corresponding to the groups. If set to NULL, a palette is generated using \code{rainbow()} based on the number of unique groups.}

\item{title}{A string specifying the file name of the PDF where the PCA plots will be saved.}

\item{ellipse}{Logical. If TRUE, a 95\% confidence ellipse is drawn on the PCA individuals plot. Default is FALSE.}

\item{comp.num}{Numeric. The number of principal components to compute and display. Default is 2.}

\item{scale}{Character. If set to "log2", a log2 transformation is applied to the numeric cytokine measurements (excluding the grouping columns). Default is NULL.}

\item{pch.values}{A vector of plotting symbols (pch values) to be used in the PCA plots. Default is NULL.}

\item{style}{Character. If set to "3d" or "3D" and \code{comp.num} equals 3, a 3D scatter plot is generated using the plot3D package. Default is NULL.}
}
\value{
A PDF file containing the PCA plots is generated and saved.
}
\description{
This function performs Principal Component Analysis (PCA) on cytokine data and generates several types of plots,
including:
\itemize{
\item 2D PCA plots using mixOmics's \code{plotIndiv} function,
\item 3D scatter plots (if \code{style} is "3d" or "3D" and \code{comp.num} is 3) via the plot3D package,
\item Scree plots showing both individual and cumulative explained variance,
\item Loadings plots, and
\item Biplots and correlation circle plots.
}
The function optionally applies a log2 transformation to the numeric data and handles analyses based on either treatment
or stimulation groups.
}
\examples{
\dontrun{
data <- cytodata[,-c(1,4, 24)]
data.df <- filter(data, Group != "ND" & Treatment != "Unstimulated")
cyt.pca(data.df, title = "Example PCA Analysis.pdf" ,colors = c("black", "red2"),
scale = "log2", comp.num = 3, pch.values = c(16,4), style = "3D",
group.col = "Group", trt.col = "Treatment")
cyt.pca(data.df, title = "Example PCA Analysis 2.pdf" ,colors = c("black", "red2"),
scale = "log2", comp.num = 2, pch.values = c(16,4), group.col = "Group")
}
}
