% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyt.pca.R
\name{cyt.pca}
\alias{cyt.pca}
\title{Analyze data with Principal Component Analysis (PCA) for cytokines.}
\usage{
cyt.pca(
  data.df,
  colors = NULL,
  title,
  ellipse = FALSE,
  comp.num = 2,
  scale = NULL,
  pch.values = NULL,
  style = NULL
)
}
\arguments{
\item{colors}{Vector of colors to be set, list of colors to be set equal to the number of groups or treatments. Default set to NULL and will result in random colors.}

\item{title}{Title of pdf file to be saved with figures.}

\item{ellipse}{95\% confidence region. Default set to FALSE.}

\item{comp.num}{Number of components to be used in PCA. Default set to 2.}

\item{scale}{Scale the data. Default set to NULL. Uses log2 transformation.}

\item{pch.values}{Vector of plotting characters to be used. Default set to NULL.}

\item{style}{Use "3d" or "3D" to generate a 3D plot. Default set to NULL.}

\item{x.df}{A data frame with cytokine data.}
}
\value{
A pdf file with PCA plots (2D and 3D), loadings plots, biplots, and correlation circle plots.
}
\description{
Analyze data with Principal Component Analysis (PCA) for cytokines.
}
\examples{
data <- cytodata[,-c(1,4)]
data.df <- filter(data, Group != "ND" & Treatment != "Unstimulated")
data.df <- data.df[,-22]
cyt.pca(data.df, title = "PCA_Example_Analysis.pdf" ,colors = c("black", "red2"), scale = "log2", comp.num = 3, pch.values = c(16,4), style = "3D")
cyt.pca(data.df, title = "PCA_Example_Analysis2.pdf" ,colors = c("black", "red2"), scale = "log2", comp.num = 2, ellipse = TRUE, pch.values = c(16,4))
}
