% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyt.plsda.R
\name{cyt.plsda}
\alias{cyt.plsda}
\title{Analyze data with Sparse Partial Least Squares Discriminant Analysis (sPLS-DA).}
\usage{
cyt.plsda(
  x.df,
  group.col,
  trt.col = NULL,
  colors = NULL,
  title,
  ellipse = FALSE,
  bg = FALSE,
  conf.mat = FALSE,
  var.num,
  cv.opt = NULL,
  fold.num = 5,
  scale = NULL,
  comp.num = 2,
  pch.values,
  style = NULL
)
}
\arguments{
\item{x.df}{A matrix or data frame of variables.}

\item{group.col}{String value of the column name for groups. Default set to NULL}

\item{trt.col}{String value of the column name for treatments.Default set to NULL}

\item{colors}{Vector of colors to be set, list of colors to be set equal to the number of groups or treatments. Default set to NULL and will result in random colors.}

\item{title}{Title of the PDF file to be saved with figures.}

\item{ellipse}{If ellipse should be drawn in the figures or not. Default set to FALSE.}

\item{bg}{If prediction background should be drawn in the figures or not. Default set to FALSE.}

\item{conf.mat}{If confusion matrix of the classifications be printed at the end or not. Default set to FALSE.}

\item{var.num}{Number of variables to be used in PLS-DA.}

\item{cv.opt}{Option for "loocv" or "Mfold" cross-validation methods.}

\item{fold.num}{Number of folds if cv.opt equals "Mfold".}

\item{scale}{Option to transform data using log2. Default set to NULL.}

\item{comp.num}{Number of components to be calculated. Default set to 2.}

\item{pch.values}{A vector of integers specifying the plotting characters to be used in the plot.}

\item{style}{A character input to have a 3D plot using '3D' or '3d'. Default set to NULL.}
}
\value{
A PDF file consisting of classification figures, component figures including Variable of importance in projection (VIP) scores, and
classification of groups with VIP scores greater than 1.
}
\description{
This function conducts Sparse Partial Least Squares Discriminant Analysis based on the groups and treatment inputed. The rest of columns will be assumed as continuous variables to be used in analysis.
The function supports log2 transformation of the data and raw data analysis. The function also supports cross-validation methods such as Leave-One-Out Cross-Validation (LOOCV) and Mfold Cross-Validation.
}
\examples{
\dontrun{
cyt.plsda(x.df, title = "Example PLS-DA Analysis.pdf", bg = TRUE, conf.mat = TRUE, scale = "log2",
var.num = 25, cv.opt = "loocv", comp.num = 2, colors = c("black", "purple", "red2"),
pch.values = c(16,4,3))
}
}
