% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyt_pls.R
\name{cyt.plsda}
\alias{cyt.plsda}
\title{Analyze data with PLS-DA method.}
\usage{
cyt.plsda(
  x.df,
  colors = NULL,
  title,
  ellipse = FALSE,
  bg = FALSE,
  conf.mat = FALSE
)
}
\arguments{
\item{x.df}{A matrix or data frame of variables.}

\item{colors}{Vector of colors to be set, list of colors to be set. Default set to NULL and will result in random colors.}

\item{title}{Title of the PDF file to be saved.}

\item{ellipse}{If ellipse should be drawn in the figures or not. Default set to FALSE.}

\item{bg}{If prediction background should be drawn in the figures or not. Default set to FALSE.}

\item{conf.mat}{If confusion matrix of the classifications be printed at the end or not. Default set to FALSE.}
}
\value{
A PDF file consisting of classification figures, component figures including Variable of importance in projection (VIP) scores, and
classification of groups with VIP scores greater than 1.
}
\description{
This function takes a matrix or data frame of variables where the first two columns must consist of "Group"
and "Stimulation" or "Group" and "Treatment". The rest of columns will be cytokines that will be analyzed.
The function also uses log2 transformation of the data, so the data that is inputed, must be raw data without
any transformation.
}
\details{
This function takes a data frame assuming that the first two columns consists
of groups and stimulation of patients and using the rest of the columns as
components for the PLS-DA analysis.
}
\examples{
cyt.plsda(data.df, colors = c("black", "purple"), title = "Example Analysis.pdf", bg = TRUE, conf.mat = TRUE)
cyt.plsda(data.df, colors = c("black", "purple"), title = "Example Analysis.pdf", ellipse = TRUE, conf.mat = TRUE)
cyt.plsda(data.df, colors = c("black", "purple"), title = "Example Analysis.pdf", conf.mat = FALSE)
cyt.plsda(data.df, ellipse = TRUE, title = "Example Analysis.pdf", conf.mat = TRUE)
}
