% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyt.volc.R
\name{cyt.volc}
\alias{cyt.volc}
\title{Volcano Plot}
\usage{
cyt.volc(
  data,
  group_col,
  cond1 = NULL,
  cond2 = NULL,
  fold_change_thresh = 2,
  p_value_thresh = 0.05,
  top_labels = 10
)
}
\arguments{
\item{data}{A matrix or data frame containing the data to be analyzed.}

\item{group_col}{A character string specifying the column name used for comparisons (e.g., group, treatment, or stimulation).}

\item{cond1}{A character string specifying the name of the first condition for comparison. Default is \code{NULL}.}

\item{cond2}{A character string specifying the name of the second condition for comparison. Default is \code{NULL}.}

\item{fold_change_thresh}{A numeric threshold for the fold change. Default is \code{2}.}

\item{p_value_thresh}{A numeric threshold for the p-value. Default is \code{0.05}.}

\item{top_labels}{An integer specifying the number of top variables to label on the plot. Default is \code{10}.}
}
\value{
A list of volcano plots (as \code{ggplot} objects) for each pairwise comparison. Additionally, the function prints the data frame
used for plotting (excluding the significance column) from the final comparison.
}
\description{
This function subsets the numeric columns from the input data and compares them based on a selected grouping column.
It computes the fold changes (as the ratio of means) and associated p-values (using two-sample t-tests) for each numeric variable
between two groups. The results are log2-transformed (for fold change) and -log10-transformed (for p-values) to generate a volcano plot.
}
\note{
If \code{cond1} and \code{cond2} are not provided, the function automatically generates all possible pairwise combinations of groups
from the specified \code{group_col} for comparisons.
}
\examples{
\dontrun{
cyt.volc(cytodata, group_col = "Group")
cyt.volc(cytodata, group_col = "Group", fold_change_thresh = 2, top_labels = 15)
}
}
