% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyt_bp2.R
\name{cyt_bp2}
\alias{cyt_bp2}
\title{Boxplot Function Enhanced for Specific Group Comparisons.}
\usage{
cyt_bp2(data, pdf_title, mf_row = c(1, 1), scale = NULL, y_lim = NULL)
}
\arguments{
\item{data}{A matrix or data frame of raw data.}

\item{pdf_title}{A string representing the title
(and filename) of the PDF file.}

\item{mf_row}{A numeric vector of length two specifying the layout
(rows and columns) for the plots on each page.
Defaults to c(1, 1).}

\item{scale}{Transformation option for continuous variables.
Options are NULL (default) and "log2". When set to "log2",
numeric columns are transformed using the log2 function.}

\item{y_lim}{An optional numeric vector defining the y-axis l
imits for the plots.}
}
\value{
A PDF file containing the boxplots.
}
\description{
This function generates a PDF file containing boxplots for each combination
of numeric and factor variables in the provided data. It first converts
any character columns to factors and checks that the data contains at
least one numeric and one factor column. If the scale argument is set to
"log2", all numeric columns are log2-transformed. The function then
creates boxplots using ggplot2 for each numeric variable grouped by
each factor variable.
}
\examples{
# Loading data
data_df <- ExampleData1[, -c(3, 4:27)]
data_df <- dplyr::filter(data_df, Group == "T2D", Treatment == "Unstimulated")
cyt_bp2(data_df, pdf_title = "boxplot2_test2.pdf", scale = "log2")

}
