% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyt_ttest.R
\name{cyt_ttest}
\alias{cyt_ttest}
\title{Two Sample T-test Comparisons.}
\usage{
cyt_ttest(data, scale = NULL)
}
\arguments{
\item{data}{A matrix or data frame containing continuous variables and
categorical variables.}

\item{scale}{A character specifying a transformation for continuous
variables. Options are \code{NULL} (default) and \code{"log2"}. When
\code{scale = "log2"}, a log2 transformation is applied and a two-sample
t-test is used; when \code{scale} is \code{NULL}, a Mann-Whitney U test is
performed.}
}
\value{
A list of p-values from the statistical tests for each
combination of a continuous outcome and a categorical predictor is returned.
}
\description{
This function performs pairwise comparisons between two groups
for each combination of a categorical predictor (with exactly two levels)
and a continuous outcome variable. It first converts any character
variables in \code{data} to factors and applies a log2 transformation to
the continuous variables if specified. Depending on the value of
\code{scale}, the function conducts either a two-sample t-test or a
Mann-Whitney U test and prints the resulting p-values. An error is thrown
if a categorical variable does not have exactly two levels.
}
\examples{
data_df <- ExampleData1[, -c(3)]
data_df <- dplyr::filter(data_df, Group != "ND", Treatment != "Unstimulated")
# Two sample T-test
cyt_ttest(data_df[, c(1, 2, 5:6)], scale = "log2")
# Mann-Whitney U Test
cyt_ttest(data_df[, c(1, 2, 5:6)])
}
