% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytodata}
\alias{cytodata}
\title{Cytokine Profiling Data}
\format{
\subsection{\code{cytodata}}{

A data frame with 297 rows and 29 columns:
\describe{
\item{Group}{Group assigned to the subjects.}
\item{Stimulation}{Stimulation recieved by subjects.}
\item{Treatment}{Treatment recieved by subjects.}
}
}
}
\source{
{Example data put together for cytokine profiling.}
}
\usage{
cytodata
}
\description{
Contains observed values of cytokines and their respective treatment and groups.
}
\examples{
data(cytodata)

}
\keyword{datasets}
