% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytProfile-package.R
\docType{package}
\name{CytoProfile-package}
\alias{CytoProfile}
\alias{CytoProfile-package}
\title{CytoProfile: Cytokine Profiling Analysis Tool}
\description{
This tool will help in conducting quality control using biological meaningful cutoff on raw measured values of cytokines. Test on distributional symmetry to suggest the adopt of transformation. Exploratory analysis including summary statistics, enriched boxplot, and barplots. Univariate analysis and Multivariate analysis options are further available to dive deeper into the cytokine profiling.
}
\author{
\strong{Maintainer}: Shubh Saraswat \email{shubh.saraswat00@gmail.com} (\href{https://orcid.org/0009-0009-2359-1484}{ORCID})

Authors:
\itemize{
  \item Xiaohua Douglas Zhang \email{douglas.zhang@uky.edu}
}

}
\keyword{internal}
