% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProfile-package.R
\docType{package}
\name{CytoProfile-package}
\alias{CytoProfile}
\alias{CytoProfile-package}
\title{CytoProfile: Cytokine Profiling Analysis Tool}
\description{
Provides comprehensive cytokine profiling analysis through quality control using biologically meaningful cutoffs on raw cytokine measurements and by testing for distributional symmetry to recommend appropriate transformations. Offers exploratory data analysis with summary statistics, enhanced boxplots, and barplots, along with univariate and multivariate analytical capabilities for in-depth cytokine profiling such as PCA, sPLS-DA, Random Forest and XGBoost.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/saraswatsh/CytoProfile}
  \item \url{https://saraswatsh.github.io/CytoProfile/}
}

}
\author{
\strong{Maintainer}: Shubh Saraswat \email{shubh.saraswat00@gmail.com} (\href{https://orcid.org/0009-0009-2359-1484}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Xiaohua Douglas Zhang \email{douglas.zhang@uky.edu} (\href{https://orcid.org/0000-0002-2486-7931}{ORCID})
}

}
\keyword{internal}
